Fix lastModified in Cipher
Last modified date should be sent by the LibrePass Application.

Now because `lastModified` date in the server database is provided by the application. It may be other than the current time. So added a new field `lastServerSync` field that have date of the update timestamp in the server database, to workaround problems related to the older `lastModified` date. For example, if the user updates the cipher when it is offline and sends it later to the server.