Refactor server errors
Moved exception to the server module and created sealed class name ServerException that contains all exceptions.

Added error field to the ServerError enum that contains a string representation of the error.

Added a message in ResponseError that is returned by some "basic errors" such as InvalidBody and some other errors.